+++ src/main/java/de/muspellheim/commons/fx/concurrent/DispatchQueue.java
@@ -15,13 +15,41 @@ import javafx.application.*;
  * <p>
  * The <code>application</code> queue run task in JavaFX application thread. The <code>background</code> queue runs task
  * in non-UI thread.
+ * <p>
+ * In <em>testing mode</em> all queues run in current thread, so you can test synchronously.
  */
 public final class DispatchQueue {
 
     private static Executor application = new ApplicationExecutor();
     private static Executor background = new QueuedExecutor("Background Queue Worker");
+    private static Executor test = new TestingExecutor();
+
+    private static boolean testing;
 
     private DispatchQueue() {
+    }
+
+    /**
+     * Check if queue runs in testing mode.
+     *
+     * @return {@code true} if queue runs in testing mode, {@code false} otherwise
+     */
+    public static boolean isTesting() {
+        return testing;
+    }
+
+    /**
+     * Activate testing mode of queue runs.
+     */
+    public static void setTesting() {
+        testing = true;
+    }
+
+    /**
+     * Deactivate testing mode of queue runs.
+     */
+    public static void unsetTesting() {
+        testing = false;
     }
 
     /**
@@ -30,7 +58,7 @@ public final class DispatchQueue {
      * @return the application queue executor
      */
     public static Executor application() {
-        return application;
+        return testing ? test : application;
     }
 
     /**
@@ -39,7 +67,7 @@ public final class DispatchQueue {
      * @return the background queue executor
      */
     public static Executor background() {
-        return background;
+        return testing ? test : background;
     }
 
     /**
@@ -56,6 +84,15 @@ public final class DispatchQueue {
         @Override
         public void execute(Runnable command) {
             Platform.runLater(command);
+        }
+
+    }
+
+    private static class TestingExecutor implements Executor {
+
+        @Override
+        public void execute(Runnable command) {
+            command.run();
         }
 
     }